refactor(frontend): split playtest vram runtime
